Refactor github-model, extract classes for auth and PRs
